+++ test/blackbox-tests/test-cases/mwe-dune-duplicate-dialect.t/duniverse/dune-configurator.2.7.1/test/dialects.t/run.t
@@ -1,0 +1,8 @@
+  $ dune build --root bad1
+  Entering directory 'bad1'
+  File "dune-project", line 9, characters 1-74:
+   9 |  (name d)
+  10 |  (implementation (extension foo2))
+  11 |  (interface (extension bar2)))
+  Error: dialect "d" is already defined
+  [1]