+++ test/blackbox-tests/test-cases/mwe-dune-duplicate-dialect.t/run.t
@@ -1,0 +1,4 @@
+This test should demonstrate that cram tests inside the vendor dir have their
+directories marked as data only
+
+  $ dune build